[style]: 계산하고자 하는 두 값 중 하나만 double형으로 형변환

+++ src/main/java/lotto/domain/Money.java
@@ -24,7 +24,7 @@ public class Money {
         if (divider.money <= 0) {
             throw BadDividerException.getInstance();
         }
-        return ((double) money) / ((double) divider.money);
+        return (double) money / divider.money;
     }
 
     @Override
